# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.Category:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Drama
        type: string
    type: object
  model.CategoryMovieCount:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      total_movies:
        type: integer
    type: object
...
  model.FavoriteRequest:
    properties:
      project_id:
        type: integer
    required:
    - project_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Ozinse Video Platform Backend
  title: Ozinse API
  version: "1.0"
paths:
...
  /api/categories:
    get:
      consumes:
      - application/json
      description: Returns all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category. Requires administrator role.
      parameters:
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create category (Admin only)
      tags:
      - categories
  /api/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a category by ID. Requires administrator role.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete category (Admin only)
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Returns a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Updates an existing category by ID. Requires administrator role.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update category (Admin only)
      tags:
      - categories
  /api/categories/movie-count:
    get:
      description: Returns a list of categories along with the total count of associated
        projects. Requires administrator role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CategoryMovieCount'
            type: array
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get movie count per category (Admin only)
      tags:
      - categories
...
  /api/users/favorites:
    post:
      consumes:
      - application/json
      description: Links a project to the authenticated user's profile as a favorite
        item
      parameters:
      - description: Favorite Payload (project_id)
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/model.FavoriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project added to favorites successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add project to favorites
      tags:
      - favorites
  /api/users/favorites/{project_id}:
    delete:
      consumes:
      - application/json
      description: Unlinks a project from the authenticated user's profile by project
        ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project removed from favorites successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid project ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found in favorites
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove project from favorites
      tags:
      - favorites
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"